graphql/requirements.txt
alembic==1.8.0
anyio==3.6.1
asyncpg==0.26.0
atomicwrites==1.4.1
attrs==21.4.0
backports.cached-property==1.0.2
click==8.1.3
colorama==0.4.5
coverage==6.4.1
fastapi==0.78.0
graphql-core==3.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
strawberry-graphql==0.117.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
watchfiles==0.15.0
websockets==10.3
zipp==3.8.0
